requirements.txt
decorator==4.0.4
django-appconf==1.0.1
django-bootstrap3==6.2.2
django-user-accounts==1.2.0
ipython==4.0.0
ipython-genutils==0.1.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
ptyprocess==0.5
pytz==2015.6
simplegeneric==0.8.1
six==1.10.0
traitlets==4.0.0
wsgiref==0.1.2
